refactoring: 기능 별 hooks로 코드 분리
