+++ client/src/hooks/index.tsx
@@ -1,1 +1,2 @@
-export { useApi } from './useApi';
+export { useLogin } from "./useLogin";
+export { useSignup } from "./useSignup";
